backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlmodel>=0.0.33
pyjwt>=2.9.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
cohere>=5.20.0
python-dateutil>=2.8.2
